Added Fixed Voronoi.cpp
This version of Voronoi.cpp includes a fix for a small bug that wasn't generating the map correctly. For each position, it has to search for the nearest site/center. In the buggy version, it would change that position to a wall/not wall on the same map that it was doing the searching. This meant that for later positions, the walls would get in the way of the search for the nearest site/center. The fix really only changes the frequency of walls in the map.